Fix T298351.
The changes introduced in 852683a incorrectly tried to usurp every account that did not previously log in through OAuth, and would fail if OAuthMigrateUsersByUsername was set to false. This is of course incorrect, since an account only needs to be usurped when it already exists on the wiki. This change fixes that and closes T298351.

Change-Id: Ie0a828e368d0e06eb815dc942770711411802db3