Add support for PluggableAuth 6.0
This commit adds support for PluggableAuth 6.0, as well as support for multiple authentication providers. Furthermore, it refactors part of the codebase and adds types to most functions and parameters. Additionally, a maintenance script is provided to migrate from pre-6.0 to this version.

To configure the remotes, you can use the $wgPluggableAuth_Config configuration parameter. For example:

$wgPluggableAuth_Config = [
    'mediawiki-wikibase-accounts' => [
        'plugin' => 'WSOAuth',
        'data' => [
            'type' => 'mediawiki',
            'uri' => 'https://example.org/index.php?title=Special:OAuth',
            'clientId' => 'omitted',
            'clientSecret' => 'omitted'
        ]
    ],
    'facebook' => [
        'plugin' => 'WSOAuth',
        'data' => [
            'type' => 'facebook',
            'clientId' => 'omitted',
            'clientSecret' => 'omitted',
            'redirectUri' => 'https://example.org/index.php/Special:PluggableAuthLogin'
        ]
    ]
];

Change-Id: I44551687fc262686ea0eaf3f9e260689f46b0d57