Add option to use the real name as the account name
This commit adds an option to use the real name of a user as the local account name. It also improves the handling of the case when a username already exists. Previously, a user would be blocked from logging in when their username was already taken on the wiki and account migration was off, now it automatically appends a number to the account name. This guarantees a user can log in. Furthermore, this commit adds the option to specify certain things directly in the PluggableAuth config, instead of through a configuration variable. The configuration variable will be used as a fallback if the config option is not available in the PluggableAuth_Config array.

This closes T310340.

Change-Id: I05578d9144bcc94fcb67516e8eef3f90075ccad6